+++ apps/cold-platform-openai/src/extraction/extraction.module.ts
@@ -2,9 +2,10 @@ import { Module } from '@nestjs/common';
 import { ExtractionService } from './extraction.service';
 import { ExtractionXlsxService } from './extraction.xlsx.service';
 import { XlsLoader } from '../langchain/custom_loaders/xls.loader';
+import { ClassificationService } from './classification.service';
 
 @Module({
-  providers: [ExtractionService, ExtractionXlsxService, XlsLoader],
-  exports: [ExtractionService, ExtractionXlsxService],
+	providers: [ExtractionService, ClassificationService, ExtractionXlsxService, XlsLoader],
+	exports: [ExtractionService, ExtractionXlsxService, ClassificationService],
 })
 export class ExtractionModule {}